# faster clamp