--- a/velox/functions/prestosql/Arithmetic.h
+++ b/velox/functions/prestosql/Arithmetic.h
@@ -170,7 +170,9 @@ struct ClampFunction {
   FOLLY_ALWAYS_INLINE void
   call(TInput& result, const TInput& v, const TInput& lo, const TInput& hi) {
     VELOX_USER_CHECK_LE(lo, hi, "Lo > hi in clamp.");
-    result = std::clamp(v, lo, hi);
+    // std::clamp emits less efficient ASM
+    const TInput& a = v < lo ? lo : v;
+    result = a > hi ? hi : a;
   }
 };