Summary:
Alternative implementation of std::clamp.  It contains the same order of bounding (lower bound before upper bounding).

Produces better ASM (https://godbolt.org/z/b6YPxsoY7)

Benchmarks:

After:
```
============================================================================
[...]hmarks/basic/FeatureNormalization.cpp     relative  time/iter   iters/s
============================================================================
normalize                                                  22.88ns    43.70M
normalizeConstant                                          30.10ns    33.22M
```

Before:
```
============================================================================
[...]hmarks/basic/FeatureNormalization.cpp     relative  time/iter   iters/s
============================================================================
normalize                                                  26.81ns    37.30M
normalizeConstant                                          31.83ns    31.42M
```

Reviewed By: pedroerp

Differential Revision: D37737908